backend/requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
gunicorn==20.1.0
Werkzeug==2.2.3
flask-cors==3.0.10
